improve: more testing of create destination
move tests to their own file, and add a test case for validation failure.
